requirements.txt
# Core dependencies
anthropic>=0.39.0
pandas>=2.2.0
pydantic>=2.9.0
python-dotenv>=1.0.0

# PDF processing
PyMuPDF>=1.24.0
Pillow>=10.0.0

# Data validation
pydantic-settings>=2.0.0

# String matching for DQS
python-Levenshtein>=0.25.0

# Code quality
ruff>=0.6.0
pytest>=8.0.0
